添加 security 安全模块

+++ tyrande-security/src/main/java/org/tyrande/security/dao/jwt/JwtUserDao.java
@@ -28,5 +28,5 @@ public interface JwtUserDao {
      * @param userId 用户编号
      * @return
      */
-    List<String> getRoleList(@Param("userId") Integer userId);
+    List<String> getRoleList(@Param("userId") Long userId);
 }
